requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
openai==1.51.0
httpx==0.25.2
redis==5.0.1
requests==2.31.0
pyyaml==6.0.1
lightgbm==4.1.0
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.24.3
aiofiles==23.2.1
python-multipart==0.0.6